requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.23.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.1.0
matplotlib>=3.4.0
